+++ src/react.ts
@@ -1,3 +1,3 @@
 export { useAtom } from "./useAtom";
 export { useSignal } from "./useSignal";
-export { useSelectAtom } from "./useSelectAtom";
+export { useSelector } from "./useSelector";